requirements.txt
Flask
pandas
numpy
streamlit
mysql.connector